+++ go-backend/models/Subscribe.go
@@ -6,10 +6,10 @@ import (
 )
 
 type Subscribe struct {
-	ID      string `json:"id" gorm:"type:char(36);primaryKey"`
-	Name    string `json:"name"`
-	Balance *uint  `json:"balance"`
-	Money   *uint  `json:"money"`
+	ID      string   `json:"id" gorm:"type:char(36);primaryKey"`
+	Name    string   `json:"name"`
+	Balance *float32 `json:"balance"`
+	Money   *float32 `json:"money"`
 }
 
 // BeforeCreate 在创建记录之前生成 UUID
